fix(BonusNumberVerifier): checkRange int -> long으로 타입 변경

+++ src/main/java/lotto/verifier/BonusNumberVerifier.java
@@ -31,7 +31,7 @@ public class BonusNumberVerifier implements Verifier{
     }
 
     private void checkRange(String input){
-        int BonusNumber = Integer.parseInt(input);
+        long BonusNumber = Long.parseLong(input);
         if(BonusNumber < Constant.START_INCLUSIVE || BonusNumber > Constant.END_INCLUSIVE ){
             throw new IllegalArgumentException(ExceptionMessage.NUMBER_EACH_OUT_OF_RANGE);
         }
